added BezierMovement
