+++ Assets/Scripts/Obstacle/BezierPath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0888abd286e8cbf40a747fac1068de87
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
